[#79] Rename: 패밀리 폴더 세분화 작업
