+++ pages/family/index.tsx
@@ -1,8 +1,8 @@
 import CalendarInstance from '@/utils/date/date.utils';
 import styles from './index.module.scss';
-import FamilyDiary from '@/components/family/family-diary';
-import FamilyGrowth from '@/components/family/family-growth';
-import FamilySchedule from '@/components/family/family-schedule';
+import FamilyDiary from '@/components/family/diary/family-diary';
+import FamilyGrowth from '@/components/family/growth/family-growth';
+import FamilySchedule from '@/components/family/schedule/family-schedule';
 
 export default function Family() {
   return (
